# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Springfield AI Command Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /export/logs:
    get:
      operationId: exportLogs
      tags: [exports]
      summary: Export daily logs as Excel
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: classId
          in: query
          schema:
            type: integer
        - name: teacherId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /export/events:
    get:
      operationId: exportEvents
      tags: [exports]
      summary: Export student events as Excel
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: classId
          in: query
          schema:
            type: integer
        - name: eventType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
...